trans_multi: undo accidentally committed change to index_t.
Ditz-issue: 8f36cd6c86489cdaa7349beb2277e947a5d0558a

+++ d2/trans_multi.h
@@ -55,7 +55,7 @@ private:
 	static ale_pos _multi_decomp;
 	static ale_real _multi_improvement;
 
-	typedef unsigned short index_t;
+	typedef unsigned int index_t;
 
 	std::vector<trans_single> trans_stack;
 	std::vector<multi_coordinate> coord_stack;
